MS-8 elementary operation: mult row

+++ unit-tests/matrix/test_swap_row.h
@@ -6,11 +6,11 @@ TEST_F(swap_row, 3x3)
   matrix_t* A = matrix_new(3, 3);
   matrix_t* B = matrix_new(3, 3);
 
-  const static int elts[] = {
+  const static float elts[] = {
     1, 2, 3, 4, 5, 6, 7, 8, 9,
   };
 
-  const static int expected[] = {
+  const static float expected[] = {
     7, 8, 9, 4, 5, 6, 1, 2, 3,
   };
 
@@ -31,11 +31,11 @@ TEST_F(swap_row, 2x3)
   matrix_t* A = matrix_new(2, 3);
   matrix_t* B = matrix_new(2, 3);
 
-  const static int elts[] = {
+  const static float elts[] = {
     1, 2, 3, 4, 5, 6,
   };
 
-  const static int expected[] = {
+  const static float expected[] = {
     4, 5, 6, 1, 2, 3,
   };
